+++ rednotebook/info.py
@@ -1,5 +1,5 @@
 # -----------------------------------------------------------------------
-# Copyright (c) 2009-2015  Jendrik Seipp
+# Copyright (c) 2008-2024 Jendrik Seipp
 #
 # RedNotebook is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -28,10 +28,10 @@ if not hasattr(builtins, "_"):
 
 program_name = "RedNotebook"
 tagline = _("A Desktop Journal")
-version = "2.32"
+version = "2.33"
 author = "Jendrik Seipp"
 author_mail = "jendrikseipp@gmail.com"
-copyright_ = "Copyright (c) 2008-2022 Jendrik Seipp"
+copyright_ = "Copyright (c) 2008-2024 Jendrik Seipp"
 url = "https://rednotebook.app"
 downloads_url = "https://rednotebook.app/downloads.html"
 donation_url = "https://rednotebook.app/downloads.html"