+++ win/fetch-dictionary.py
@@ -6,6 +6,7 @@ import shutil
 
 import utils
 
+
 DIR = os.path.dirname(os.path.abspath(__file__))
 DICT_DIR = os.path.join(DIR, "dicts")